+++ app/src/main/java/com/yunarm/appstore/ftp/FtpDownLoadListener.java
@@ -1,7 +1,7 @@
 package com.yunarm.appstore.ftp;
 
 public interface FtpDownLoadListener {
-    void onDownProgress(long progress);
+    void onDownProgress(int progress);
     void onDownloadSucc(String localPath);
     void onDownloadFail( );
 }
